+++ src/Suru.Compiler/Parse/Ast/AssignmentStatement.cs
@@ -1,0 +1,7 @@
+namespace Suru.Compiler.Parse.Ast;
+
+public sealed class AssignmentStatement(string name, Expression value) : Statement
+{
+    public string Name { get; } = name;
+    public Expression Value { get; } = value;
+}