+++ src/Suru.Compiler/Parse/Ast/VariableReferenceExpression.cs
@@ -1,0 +1,6 @@
+namespace Suru.Compiler.Parse.Ast;
+
+public sealed class VariableReferenceExpression(string name) : Expression
+{
+    public string Name { get; } = name;
+}