+++ src/Suru.Compiler/Types/SuruType.cs
@@ -1,0 +1,8 @@
+namespace Suru.Compiler.Types;
+
+public enum SuruType
+{
+    Bool,
+    Int64,
+    Float64,
+}